S3.7: CodeRabbit follow-ups — CMake fail-fast + uint16_t ports
- CMake: FATAL_ERROR when SOLIDSYSLOG_OPENSSL=ON but OpenSSL not found
  (was: silent fall-through to confusing downstream errors).
- Example{Tls,Tcp,Udp}Config_GetPort now returns uint16_t instead of int,
  matching SolidSyslogEndpoint::port domain; redundant uint16_t casts at
  call sites removed.

Verified locally: debug (725 tests), debug -DSOLIDSYSLOG_OPENSSL=OFF
(612 tests), tidy, cppcheck, format all green.

+++ Example/Common/ExampleTcpConfig.h
@@ -9,7 +9,7 @@
 EXTERN_C_BEGIN
 
     const char* ExampleTcpConfig_GetHost(void);
-    int         ExampleTcpConfig_GetPort(void);
+    uint16_t    ExampleTcpConfig_GetPort(void);
     void        ExampleTcpConfig_GetEndpoint(struct SolidSyslogEndpoint * endpoint);
     uint32_t    ExampleTcpConfig_GetEndpointVersion(void);
