+++ Example/Common/ExampleUdpConfig.c
@@ -14,15 +14,15 @@ const char* ExampleUdpConfig_GetHost(void)
     return "syslog-ng";
 }
 
-int ExampleUdpConfig_GetPort(void)
+uint16_t ExampleUdpConfig_GetPort(void)
 {
-    return EXAMPLE_UDP_PORT;
+    return (uint16_t) EXAMPLE_UDP_PORT;
 }
 
 void ExampleUdpConfig_GetEndpoint(struct SolidSyslogEndpoint* endpoint)
 {
     SolidSyslogFormatter_BoundedString(endpoint->host, ExampleUdpConfig_GetHost(), SOLIDSYSLOG_MAX_HOST_SIZE);
-    endpoint->port = (uint16_t) ExampleUdpConfig_GetPort();
+    endpoint->port = ExampleUdpConfig_GetPort();
 }
 
 /* Static config — host/port never change, so version stays 0 forever and the
